Move tailscale base entity to separate module
